Replace CredsForSessCfg.addCredsOrPrompt with ConnectionPropsForSessCfg.addPropsOrPrompt
Add and/or prompt for host and port.
