+++ packages/rest/index.ts
@@ -25,7 +25,7 @@ export * from "./src/client/RestClientError";
 export * as SessConstants from "./src/session/SessConstants";
 
 export * from "./src/session/doc/ISession";
-export * from "./src/session/doc/IOptionsForAddCreds";
+export * from "./src/session/doc/IOptionsForAddConnProps";
 export * from "./src/session/AbstractSession";
 export * from "./src/session/Session";
-export * from "./src/session/CredsForSessCfg";
+export * from "./src/session/ConnectionPropsForSessCfg";
